backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
cryptography==42.0.2
python-dotenv==1.0.0
httpx==0.26.0
redis==5.0.1
prometheus-client==0.19.0
slowapi==0.1.9
email-validator==2.1.0
pymysql==1.1.0
mysql-connector-python==8.3.0
